flow_agent/pyproject.toml
[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "src"
module-root = ""

[project]
name = "flow_agent"
version = "0.1.0"
description = "Add your description here"
# readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.5",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
extract-profile-skill = "src.agents.extract_profile_skill:main"
select-push-card = "src.agents.select_push_card:main"
render-push-card = "src.agents.render_push_card:main"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
addopts = "-v -s"
testpaths = ["tests"]
